# docs/swagger.yaml
definitions:
  request.StaffLoginRequest:
    properties:
      hospital:
        example: 1
        type: integer
      password:
        example: "1234"
        type: string
      username:
        example: test
        type: string
    required:
    - hospital
    - password
    - username
    type: object
  request.StaffRequest:
    properties:
      age:
        example: 10
        type: integer
      email:
        example: admin@gmail.com
        type: string
      first_name_en:
        example: test
        type: string
      first_name_th:
        example: ทดสอบ
        type: string
      gender:
        example: M
        type: string
      hospital:
        example: 1
        type: integer
      last_name_en:
        example: tester
        type: string
      last_name_th:
        example: นามสกุล ทดสอบ
        type: string
      password:
        example: "1234"
        type: string
      username:
        example: test
        type: string
    required:
    - age
    - email
    - first_name_en
    - first_name_th
    - gender
    - hospital
    - last_name_en
    - last_name_th
    - password
    - username
    type: object
  response.CreateDataResponse:
    properties:
      message:
        example: create success
        type: string
      response_key:
        type: string
      response_message:
        type: string
    type: object
  response.HospitalModel:
    properties:
      address:
        type: string
      id:
        type: integer
      name_en:
        type: string
      name_th:
        type: string
      phone_number:
        type: string
    type: object
  response.LoginStaffModel:
    properties:
      refresh_token:
        example: admin
        minLength: 1
        type: string
      token:
        example: "1234"
        minLength: 1
        type: string
    required:
    - refresh_token
    - token
    type: object
  response.PatientSearchModel:
    properties:
      date_of_birth:
        type: string
      email:
        type: string
      first_name_en:
        type: string
      first_name_th:
        type: string
      gender:
        type: string
      last_name_en:
        type: string
      last_name_th:
        type: string
      middle_name_en:
        type: string
      middle_name_th:
        type: string
      national_id:
        type: string
      passport_id:
        type: string
      patient_hn:
        type: string
      phone_number:
        type: string
    type: object
info:
  contact: {}
paths:
  /hospital:
    get:
      description: Search Hospital
      parameters:
      - description: int valid
        in: query
        name: page
        type: integer
      - description: int valid
        in: query
        name: pageSize
        type: integer
      - description: string valid
        in: query
        name: sortField
        type: string
      - description: string valid
        in: query
        name: sortValue
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.HospitalModel'
      security:
      - Bearer: []
      summary: Get List Hospital
      tags:
      - Hospital
  /patient/search:
    get:
      description: Search Detail Patient
      parameters:
      - description: string valid
        in: query
        name: national_id
        type: string
      - description: string valid
        in: query
        name: passport_id
        type: string
      - description: string valid
        in: query
        name: first_name_en
        type: string
      - description: string valid
        in: query
        name: first_name_th
        type: string
      - description: string valid
        in: query
        name: middle_name_en
        type: string
      - description: string valid
        in: query
        name: middle_name_th
        type: string
      - description: string valid
        in: query
        name: last_name_en
        type: string
      - description: string valid
        in: query
        name: last_name_th
        type: string
      - description: string valid
        in: query
        name: date_of_birth
        type: string
      - description: string valid
        in: query
        name: phone_number
        type: string
      - description: string valid
        in: query
        name: email
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.PatientSearchModel'
      security:
      - Bearer: []
      summary: Search Detail Patient
      tags:
      - Patient
  /patient/search/{ID}:
    get:
      description: Search Patient
      parameters:
      - description: ID
        in: path
        name: ID
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.PatientSearchModel'
      security:
      - Bearer: []
      summary: Search Patient
      tags:
      - Patient
  /staff:
    post:
      description: Create Staff
      parameters:
      - description: query params
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/request.StaffRequest'
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.CreateDataResponse'
      security:
      - Bearer: []
      summary: Create Staff
      tags:
      - Staff
  /staff/login:
    post:
      description: Login Staff
      parameters:
      - description: query params
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/request.StaffLoginRequest'
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.LoginStaffModel'
      security:
      - Bearer: []
      summary: Login Staff
      tags:
      - Staff
securityDefinitions:
  Bearer:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"